requirements.txt
beautifulsoup4==4.12.2
playwright==1.40.0
python-dotenv==1.0.0
openai==1.7.2